instantback/app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
google-generativeai
pinecone-client
langchain
python-dotenv
